saving pak file extension to savegame
git-svn-id: svn://tron.homeunix.org/simutrans/simutrans/trunk@670 8aca7d54-2c30-db11-9de9-000461428c89

+++ dings/leitung2.cc
@@ -422,7 +422,7 @@ void leitung_t::laden_abschliessen()
  */
 void leitung_t::rdwr(loadsave_t *file)
 {
-	unsigned long value;
+	uint32 value;
 
 	ding_t::rdwr(file);
 	if(file->is_saving()) {
